feat: change data displaying
