+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { use } from 'react';
 
-import { CountriesList } from './components/CountriesList';
+import { DataTable } from './components/DataTable';
 
 import { getData } from '@/api/co2Data';
 import type { CO2Data } from '@/types/table';
@@ -13,7 +13,12 @@ export const App = () => {
       <h1 className="mb-3 text-center text-3xl">
         CO<sub>2</sub> Data
       </h1>
-      <CountriesList countries={data} />
+
+      <div>
+        <button>Configure Years Tables</button>
+      </div>
+
+      <DataTable countries={data} />
     </div>
   );
 };
