--- src/components/CountriesList/index.ts
@@ -1,1 +1,0 @@
-export * from './CountriesList';
