--- src/components/CountriesListHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './CountriesListHeader';
