--- src/components/CountryListRow/index.ts
@@ -1,1 +1,0 @@
-export * from './CountryListRow';
