+++ src/components/DataTableRow/index.ts
@@ -1,0 +1,1 @@
+export * from './DataTableRow';
