+++ src/components/YearsTableColumnsModal/YearsTableColumnsModal.tsx
@@ -1,0 +1,3 @@
+export const YearsTableColumnsModal = () => {
+  return <div>YearsTableColumnsModal</div>;
+};
