--- src/components/YearsTableHeader/index.ts
@@ -1,1 +1,0 @@
-export * from './YearsTableHeader';
